Allow per-service annotations
We add the 'annotations' field to the existing
vault.service.{active,standby} dictionaries which are relevant for the
active/standby vault ha services. We also add
vault.service.{nonha,internal}.annotations in order to allow per-service
annotations when using the non-ha variant.

We had to choose 'nonha' as we cannot reuse the existing
vault.service.annotations key, because that gets still applied to all
services and we do not want to break existing installations.

WIP as we need to add some more docs and maybe some more tests.